requirements.txt
torch==2.10.0+cpu
torchvision==0.15.0+cpu
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
python-multipart==0.0.6
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.0